Разбил приложение company_details на два: staff, contacts
