+++ Electronics/contacts/admin.py
@@ -2,11 +2,6 @@ from django.contrib import admin
 from .models import *
 
 
-@admin.register(Staff)
-class StaffAdmin(admin.ModelAdmin):
-    list_display = [field.name for field in Staff._meta.fields]
-
-
 @admin.register(Address)
 class AddressAdmin(admin.ModelAdmin):
     list_display = [field.name for field in Address._meta.fields]
