+++ Electronics/contacts/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class CompaniesDetailsConfig(AppConfig):
+class ContactsConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "companies_details"
+    name = "contacts"
