+++ Electronics/staff/admin.py
@@ -1,0 +1,7 @@
+from django.contrib import admin
+from .models import *
+
+
+@admin.register(Staff)
+class StaffAdmin(admin.ModelAdmin):
+    list_display = [field.name for field in Staff._meta.fields]
